requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
Brotli==1.1.0
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
css-html-js-minify==2.5.5
cssselect==1.3.0
cssselect2==0.7.0
cssutils==2.11.1
Django==5.1.2
django-appconf==1.1.0
django-celery-beat==2.7.0
django-crontab==0.7.1
django-extensions==3.2.3
django-select2==8.4.0
django-simple-history==3.7.0
django-tenants==3.7.0
django-timezone-field==7.0
djangorestframework==3.16.0
docopt==0.6.2
et_xmlfile==2.0.0
fonttools==4.55.7
frozenlist==1.5.0
htmlmin==0.1.12
idna==3.10
Jinja2==3.1.6
kombu==5.4.2
lxml==5.3.1
MarkupSafe==3.0.2
more-itertools==10.6.0
multidict==6.2.0
num2words==0.5.13
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
pillow==11.0.0
premailer==3.10.0
prompt_toolkit==3.0.48
propcache==0.3.0
psycopg2-binary
pycparser==2.22
pydot==3.0.2
pydyf==0.11.0
PyJWT==2.10.1
pyparsing==3.2.0
pyphen==0.17.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
reportlab==4.2.5
reportlib==3.4.0
requests==2.32.3
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.1
stripe==11.4.1
tinycss2==1.4.0
tinyhtml5==2.0.0
twilio==9.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
weasyprint==63.1
webencodings==0.5.1
wheel==0.45.1
yarl==1.18.3
zopfli==0.2.3.post1
